+++ src/main/java/com/ef/services/logparse/LogParseInterface.java
@@ -1,10 +1,11 @@
-package com.ef.services.log_parse;
+package com.ef.services.logparse;
 
 /**
  * Generic Interface
  * @param <T1> Return object
  * @param <T2> Input object
  */
+@FunctionalInterface
 public interface LogParseInterface <T1 extends Object, T2 extends Object> {
     public T1 parse(T2 input);
 }
